Move role management to club home
