+++ app/i18n/en/common.json
@@ -96,6 +96,7 @@
   "president": "President",
   "vicePresident": "Vice President",
   "member": "Member",
+  "manageRoles": "Manage Roles",
   "clubDirectory": "Club Directory",
   "myClubs": "My Clubs",
   "clubsDirectory": "Clubs Directory",
