+++ app/i18n/es/common.json
@@ -96,6 +96,7 @@
     "president": "Presidente",
     "vicePresident": "Vicepresidente",
     "member": "Miembro",
+    "manageRoles": "Administrar roles",
     "clubDirectory": "Directorio de clubes",
     "myClubs": "Mis Clubes",
     "clubsDirectory": "Directorio de clubes",
