+++ app/i18n/vi/common.json
@@ -96,6 +96,7 @@
     "president": "Chủ tịch",
     "vicePresident": "Phó chủ tịch",
     "member": "Thành viên",
+    "manageRoles": "Quản lý vai trò",
     "clubDirectory": "Thư mục câu lạc bộ",
     "myClubs": "Câu lạc bộ của tôi",
     "clubsDirectory": "Thư mục câu lạc bộ",
