+++ app/i18n/zh/common.json
@@ -96,6 +96,7 @@
   "president": "会长",
   "vicePresident": "副会长",
   "member": "成员",
+  "manageRoles": "管理角色",
   "clubDirectory": "俱乐部目录",
   "myClubs": "我的俱乐部",
   "clubsDirectory": "俱乐部目录",
